refactor code
add controllers and schemas folders. Transfer and import/export methods
